requirements.txt
cryptography
pycryptodome
websockets
aiohttp
aioconsole